LayoutTests/fast/loader/url-parse-1-expected.txt

The following table shows how URLs are parsed and canonicalized. The base URI for this document is file:///BASE/.

URL	href	host	pathname
file:///BASE/		/BASE/
test	file:///BASE/test		/BASE/test
/	file:///		/
/test	file:///test		/test
//	file:///		/
//test	file://test/	test	/
///	file:///		/
///test	file:///test		/test
file:	file:///BASE/		/BASE/
file:test	file:///BASE/test		/BASE/test
file:/	file:///		/
file:/test	file:///test		/test
file://	file:///		/
file://test	file://test/	test	/
file:///	file:///		/
file:///test	file:///test		/test
file://localhost	file:///		/
file://localhost/	file:///		/
file://localhost/test	file:///test		/test
http:	http:/		/
http:/	http:/		/
http://	http:/		/
http:///	http:/		/
http:////	http:/		/
http:webkit.org	http://webkit.org/	webkit.org	/
http:/webkit.org	http://webkit.org/	webkit.org	/
http://webkit.org	http://webkit.org/	webkit.org	/
http:///webkit.org	http://webkit.org/	webkit.org	/
http:////webkit.org	http://webkit.org/	webkit.org	/
http://///webkit.org	http://webkit.org/	webkit.org	/
http://////webkit.org	http://webkit.org/	webkit.org	/
http:////////////////webkit.org	http://webkit.org/	webkit.org	/
http://localhost	http://localhost/	localhost	/
http://localhost/	http://localhost/	localhost	/
http://localhost/test	http://localhost/test	localhost	/test
x-webkit:	x-webkit:		
x-webkit:test	x-webkit:test		test
x-webkit:/	x-webkit:/		/
x-webkit:/test	x-webkit:/test		/test
x-webkit://	x-webkit://		//
x-webkit://test	x-webkit://test	test	
x-webkit:///	x-webkit:///		///
x-webkit:///test	x-webkit:///test		///test
